Refactor rename lib/model/utils/adamantite_file_utils.rb to lib/model/file_utils/adamantite_file_utils.rb
